Factor IR helpers into submodules
Extract the cycle detection logic and command interpolation helpers into dedicated ir::cycle and ir::cmd_interpolate modules so src/ir.rs stays below the 400 line limit. Update BuildGraph to call the new analysis API, add tests covering command interpolation behaviour in the new module, and refresh the design docs and doctest footnotes to reflect the module layout.
